correct type of Beatmap#diff_overall

+++ src/osu/mappings/beatmap.cr
@@ -9,7 +9,7 @@ module Osu
       version:          String,
       md5:              {key: "file_md5", type: String},
       diff_size:        {type: Float32?, converter: StringConverter::Float32, getter: false, setter: false},
-      diff_overall:     {type: UInt32?, converter: StringConverter::UInt32, getter: false, setter: false},
+      diff_overall:     {type: Float32?, converter: StringConverter::Float32, getter: false, setter: false},
       diff_approach:    {type: Float32?, converter: StringConverter::Float32, getter: false, setter: false},
       diff_drain:       {type: UInt32?, converter: StringConverter::UInt32, getter: false, setter: false},
       mode:             {type: API::Mode, converter: ModeConverter},
@@ -56,7 +56,7 @@ module Osu
   record(
     Difficulty,
     size : Float32?,
-    overall : UInt32?,
+    overall : Float32?,
     approach : Float32?,
     drain : UInt32?,
     rating : Float32?
